fix: evaluate literal arithmetic as exact rationals
Keep fractional and oversized literal intermediates exact instead of
truncating divisions or falling through to runtime word arithmetic.
Reject nonintegral integer conversions and preserve typed division.

Bound rational components to 4096 bits and fold expressions whose
intermediates cannot be represented by EVM words. Preserve exponent-zero
precedence shared with CYCLOPS-1469.

+++ tests/ui/codegen/lowering/run-call/signed_integer_literals.sol
@@ -76,7 +76,7 @@ contract SignedIntegerLiterals {
     }
 
     function largePositiveDivision() external pure returns (uint256) {
-        return 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff / 2;
+        return uint256(0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff) / 2;
     }
 
     function largePositiveRightShift(uint256 shift) external pure returns (uint256) {
